requirements.txt
asgiref==3.7.2
azure-core==1.29.4
azure-storage-blob==12.17.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-storages==1.14
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
Pillow==10.0.0
pycparser==2.21
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
